修复：确保序列号在更新时使用正确的类型
- 将序列号的计算中的类型转换为 uint32_t，以避免潜在的类型问题。
- 更新了 prop_info 和 SystemProperties 中的序列号处理逻辑，以确保一致性和正确性。

+++ prop_info.cpp
@@ -48,7 +48,7 @@ prop_info::prop_info(const char* name, uint32_t namelen, uint32_t long_offset) {
   this->name[namelen] = '\0';
 
   auto error_value_len = sizeof(kLongLegacyError) - 1;
-  atomic_store_explicit(&this->serial, error_value_len << 24 | kLongFlag, memory_order_relaxed);
+  atomic_store_explicit(&this->serial, (uint32_t)(error_value_len << 24 | kLongFlag), memory_order_relaxed);
   memcpy(this->long_property.error_message, kLongLegacyError, sizeof(kLongLegacyError));
 
   this->long_property.offset = long_offset;
